Implement demand paging for flexible memory
This replaces ENABLE_FULL4G with signal-based demand paging, providing:
- Automatic memory growth: pages allocated only when accessed
- Minimal footprint: typical programs use 128KB-1MB vs 256MiB virtual
- Incremental GC: unused zero-filled chunks reclaimed via madvise
- Add 4GB size limit check in memory_new() to protect bitmap bounds
- Improve signal handler documentation (POSIX.1-2008 safety)
- Cap Emscripten builds to 512MB (no mmap support)
- Improve non-MMAP fallback documentation

+++ src/devices/virtio-blk.c
@@ -88,7 +88,8 @@ static void virtio_blk_set_fail(virtio_blk_state_t *vblk)
 static inline uint32_t vblk_preprocess(virtio_blk_state_t *vblk UNUSED,
                                        uint32_t addr)
 {
-    if ((addr >= MEM_SIZE) || (addr & 0b11)) {
+    /* Cast to uint64_t for comparison since MEM_SIZE may be 4GB */
+    if (((uint64_t) addr >= MEM_SIZE) || (addr & 0b11)) {
         virtio_blk_set_fail(vblk);
         return 0;
     }
